requirements.txt
spacy
floret
pandas
spacy-huggingface-hub